backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Django API Backend"
authors = ["lupoyes <arthur.roessler@icloud.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.5"
djangorestframework = "^3.16.1"
django-cors-headers = "^4.7.0"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
whitenoise = "^6.9.0"
dj-database-url = "^3.0.1"
pillow = "^11.3.0"
sqlparse = "^0.5.3"
asgiref = "^3.9.1"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
flake8 = "^6.0.0"
pytest = "^7.0.0"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"